*: Rename pkg/chargeback to pkg/operator
