+++ pkg/operator/aws_usage_hive.go
