+++ pkg/operator/generate_report.go
