+++ pkg/operator/http.go
@@ -24,8 +24,8 @@ import (
 
 	api "github.com/operator-framework/operator-metering/pkg/apis/metering/v1alpha1"
 	cbutil "github.com/operator-framework/operator-metering/pkg/apis/metering/v1alpha1/util"
-	"github.com/operator-framework/operator-metering/pkg/chargeback/prestostore"
 	listers "github.com/operator-framework/operator-metering/pkg/generated/listers/metering/v1alpha1"
+	"github.com/operator-framework/operator-metering/pkg/operator/prestostore"
 	"github.com/operator-framework/operator-metering/pkg/presto"
 	"github.com/operator-framework/operator-metering/pkg/util/orderedmap"
 )
