+++ pkg/operator/query.go
