+++ pkg/operator/reports.go
