+++ pkg/operator/templates.go
