+++ pkg/operator/util.go
