+++ pkg/operator/util_test.go
